1. Add preference to force off Game Mode 2. Manually forcing Game Mode on/off/automatic when BetterGameMode itself forced it on now overrides the user default 3. User defaults changes take effect immediately, rather than the next time something changes
Yes, I'm aware that the notification says "Quitting the app..." when a app was running that forced on Game Mode was removed from the force-on list, but I don't really have a good way to differentiate here. I don't think it will happen enough that I need to worry about this.